implementing mapstringmapstringdouble

+++ src/packagedmaterial.cc
@@ -15,13 +15,13 @@ PackagedMaterial::Ptr PackagedMaterial::Create(Agent* creator, double quantity,
                              PackagedMaterial::package quality) {
   if (qualids_.count(quality) == 0) {
     qualids_[quality] = next_qualid_++;
-    std::map<std::string, std::map<std::string, int>> mapmap = quality.second;
-    std::map<std::string, int> insidemap; 
-    std::vector<int> insideval;
+    std::map<std::string, std::map<std::string, double>> mapmap = quality.second;
+    std::map<std::string, double> insidemap; 
+    std::vector<double> insideval;
     std::vector<std::string> columnname;
-    for (std::map<std::string, std::map<std::string, int>>::iterator it = mapmap.begin();it!=mapmap.end(); ++it){
+    for (std::map<std::string, std::map<std::string, double>>::iterator it = mapmap.begin();it!=mapmap.end(); ++it){
       insidemap = it->second; 
-      for (std::map<std::string, int>::iterator itt = insidemap.begin();itt!=insidemap.end();++itt){
+      for (std::map<std::string, double>::iterator itt = insidemap.begin();itt!=insidemap.end();++itt){
         insideval.push_back(itt->second);
         columnname.push_back(it->first + itt->first); 
       }
